requirements.txt
fastapi
streamlit
tensorflow
azure-cosmos
azure-identity
azure-storage-blob
python-dotenv
uvicorn
black
ruff
types-requests
pandas
pandas-stubs
mypy
pytest
pre-commit
authlib